refactor(reposicao): extrai types/helpers de AdminReposicaoNegociacaoParalela
Primeiro corte do god-component AdminReposicaoNegociacaoParalela (1199
LoC). Passo seguro (Codex order #1): extrai as folhas puras pra
src/components/reposicao/negociacaoParalela/, zero mudança de lógica:

- types.ts   — StatusSugestao, Categoria, OrdenacaoKey, Sugestao,
               RankingRow + consts (EMPRESA, CATEGORIAS, STATUS_LIST,
               ORDENACOES)
- helpers.ts — formatBRL, formatPerc, categoriaBadgeClass,
               categoriaLabel, statusBadgeClass, statusLabel,
               percPromoBadgeClass, lastDayOfNextMonth (todas puras,
               retornam string/classe, sem JSX)

Página: 1199 → 1053 LoC. Os 2 módulos entram no strict (350); a página
já era strict-clean e continua.

Verificado: tsc 0, typecheck:strict 0, eslint-unused 0, build OK, 423
testes passando.
