+++ tsconfig.strict.json
@@ -366,6 +366,8 @@
     "src/components/reposicao/routePlanner/priority.ts",
     "src/components/reposicao/routePlanner/constants.ts",
     "src/components/des/PosicaoAtualTab.tsx",
-    "src/components/des/HistoricoTab.tsx"
+    "src/components/des/HistoricoTab.tsx",
+    "src/components/reposicao/negociacaoParalela/types.ts",
+    "src/components/reposicao/negociacaoParalela/helpers.ts"
   ]
 }
